check for kubectl before running completion script

+++ .bash_completion.d/kubectl.sh
@@ -1,8 +1,10 @@
 #!/bin/bash
 
-TMPKCB=/tmp/KCB${RANDOM}
-
-kubectl completion bash > $TMPKCB
-
-. $TMPKCB
-rm $TMPKCB
+if command -v kubectl; then
+  TMPKCB=/tmp/KCB${RANDOM}
+  
+  kubectl completion bash > $TMPKCB
+  
+  . $TMPKCB
+  rm $TMPKCB
+fi
